refactor(test-data): limit mock HTTP functionallity (#94)
The HTTP mock is just that. It only mocks the HTTP calls. All data setup, etc. is moved into the tests that need to data, making everything far less dependent on some global setup.

This also means that each test needs to potentially do more set up, but that is a good trade-off over trying to not break tests with global data changes.