feat(teas): move the purchase link and price to the tea
The use case here is that each tea is going to be from a specific company anyhow. One company's sencha is not going to be of the same quality as another's so there is no point in them sharing a rating, etc.

As a result, there is also no reason for them to share purchase links.